Extract util functions into standalone unit.
